requirements.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
DateTime==5.3
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==1.5.3
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
zope.interface==6.1
